ML_Engineering_Projects/Area_1_ML_System_Design/01_Semantic_Search_Engine/pyproject.toml
[project]
name = "semantic-search-engine"
version = "0.1.0"
description = "Dense vector search engine using sentence embeddings and FAISS/ChromaDB"
requires-python = ">=3.11"
dependencies = [
    "sentence-transformers>=2.7",
    "faiss-cpu>=1.8",
    "chromadb>=0.5",
    "datasets>=2.20",
    "torch>=2.3",
    "transformers>=4.41",
    "numpy>=1.26",
    "pandas>=2.2",
    "tqdm>=4.66",
    "rich>=13.7",
    "python-dotenv>=1.0",
    "umap-learn>=0.5",
    "matplotlib>=3.9",
    "pytest>=8.2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = [
    "jupyter>=1.0",
    "ipykernel>=6.29",
    "pytest-cov>=5.0",
    "ruff>=0.4",
]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
line-length = 88
